get float rand vals

+++ migSimCode/agents.h
@@ -28,7 +28,7 @@ using Ann = Network<float,
 >;
 
 // pick rand node weights
-std::uniform_real_distribution<double> dist(-1.f, 1.f);
+std::uniform_real_distribution<float> dist(-1.f, 1.f);
 
 // pick rand position
 // std::uniform_real_distribution<double> pos(0.0, 200.0);
